Limit the number of records a user can check out in a week.
* Uses cookies to accomplish this. Not the most secure, as a motivated user could
just clear them, but enough to prevent a non-malicious actor from burning through them.
* In order to use cookies, a few conditions need to be met: requests from the client
need to pass the cookies back -- see changes in Client.js, and Rails has to include
cookies modules. I took the easy way and switched off API mode. There's definitely
a way to do this with a smaller footprint. Feel free to take the initiative if
you find it.